requirements.txt
openai>=1.7.0
asyncio>=3.4.3
python-dotenv>=0.19.0
pyperclip>=1.8.2
requests>=2.25.1
